generalise compress algo

+++ src/vec2.rs
@@ -1,5 +1,5 @@
 struct Vec2 {
-    pub x: i16,
-    pub y: i16,
+    pub x: i8,
+    pub y: i8,
 }
